backend/requirements.txt
flask
pytest-flask
pytest-cov